[sensorfwd] Use float values for XYZ data. Fixes JB#60407
Accelerometer values have granularity of 1 mG even if underlying
sensors would report values with higher accuracy. This happens
implicitly due to internal to sensorfwd storage using integers
for storing xyz vectors.

Switch from integers to single precision floating point numbers.

Applications that use qtsensors API should work without any changes.

Applications that use sensors with xyz data over custom unix socket
io channel need to be adapted to data type change.

+++ datatypes/genericdata.h
@@ -65,11 +65,11 @@ public:
      * @param y Y coordinate.
      * @param z Z coordinate.
      */
-    TimedXyzData(const quint64& timestamp, int x, int y, int z) : TimedData(timestamp), x_(x), y_(y), z_(z) {}
+    TimedXyzData(const quint64& timestamp, float x, float y, float z) : TimedData(timestamp), x_(x), y_(y), z_(z) {}
 
-    int x_; /**< X value */
-    int y_; /**< Y value */
-    int z_; /**< Z value */
+    float x_; /**< X value */
+    float y_; /**< Y value */
+    float z_; /**< Z value */
 };
 Q_DECLARE_METATYPE ( TimedXyzData )
